Fix the implementation of the layouts, we should send to the theme renderer associated with an element (the theme for the widget for example) instead of the ender element's renderer Pass correctly the coordinates for the mouse move event Given that we should always return a renderer on the eon_element's implementation, make the wrapper return something, having a wrapped ender or not More fixes

+++ src/include/private/eon_input_private.h
@@ -26,7 +26,7 @@ Eon_Input * eon_input_new(void);
 
 Eon_Input_State * eon_input_state_new(Eon_Input *i, Ender_Element *e, Eon_Input_Element_Get element_get);
 void eon_input_state_feed_mouse_move(Eon_Input_State *eis,
-		unsigned int x, unsigned int y);
+		double x, double y);
 void eon_input_state_feed_mouse_in(Eon_Input_State *eis);
 void eon_input_state_feed_mouse_out(Eon_Input_State *eis);
 void eon_input_state_feed_mouse_down(Eon_Input_State *eis);
